Bug 1637926 AET clients_daily draft
See https://bugzilla.mozilla.org/show_bug.cgi?id=1637926

The first draft here defines two separate tables:

- `moz-fx-data-shared-prod.firefox_accounts_derived.aet_desktop_clients_daily_v0`
- `moz-fx-data-shared-prod.firefox_accounts_derived.aet_log_users_daily_v0`

These tables have been created with data from 2020-09-01 through 2020-09-07
for the sake of discussion. The `_v0` is meant to convey these are very much
WIP and are not expected to be productionized as-is.

These daily tables are going to rely on an early stage of ETL where we track
associations between ecosystem_user_ids, and associations of ecosystem_client_id
to ecosystem_user_id. We do not yet make any attempt here to join in canonical
IDs.

These tables also currently show exact aggregations. We will need to institute
bucketing of some sort, perhaps by having a first aggregation over all clients
to get quantiles for each metric, and then replacing the per-client aggregations
with the associated quantile. This could probably use DS help as we refine
the methodology.

This first draft is primarily intended to express a sense of what fields we
would present in an `ecosystem_clients_daily` view.